Look up Profiles from the query cache when using select_related().
To solve this, the columns now do a `select_related()` on the profile
field, and

`User.get_profile()` has been updated to reuse the same caching variable
that's populated for `select_related()` queries and ForeignKey
relations. This means that if a profile is ever included using
`select_related()` or looked up using the field relation, we can call
`get_profile()` without performing an additional lookup.

This is a simple backport from the Profile code in
4a9dc80efe8d3d2bc8ba96a923971f804fab2ebc on release-3.0.x.