Allow passing relative paths to rb-site.
Unless an absolute path is provided to rb-site, an attempt will be made
to look up the value as a site name in a global listing. This means
relative paths are treated as site names as well, which can be
unexpected.

Now, if the path separator is found in the provided name, it's treated
as a path and not a site name.

Testing Done:
Tested absolute, relative, and name-based sites.

Reviewed at https://reviews.reviewboard.org/r/8905/