Prevent links to bugs in e-mails for review requests without a repository.
The e-mail templates for review request e-mails made a bad assumption
that the review request in question had a repository. This is from the
time when all review requests required a repository. The assumption
caused a check for bug link building to go down the wrong code path,
resulting in every bug having a link to "None" (or, in the case of text
e-mails, having a "None" simply appear below the bug number).

The templates now check for the existence of the repository first,
solving this. The template tag usage has also been modernized in these
sections.

Testing Done:
Tested HTML and plain-text e-mails for review requests without a
repository, with a repository but no bug tracker, and with a repository with
a bug tracker.

Reviewed at https://reviews.reviewboard.org/r/8901/