Fix the display of bug links in the dashboard when there's no repository.
When a repository is not configured with a bug tracker, or a review
request is not associated with a repository, the bug links shown in the
dashboard still appear as clickable links. When there's multiple bugs,
they appear as one big link. This is because the row is set to be
clickable, and the column wasn't specifying it wanted to link to
something itself.

Now the column turns linking on but doesn't return any global links,
ensuring only the specific bug links within the cell are clickable.
There's also some styling fixes to make sure the links look correct and
that bugs don't prematurely wrap when there's no link to the bug.

Testing Done:
Tested links and cursor for the bugs cells for review requests without
a repository, review requests with a repository and no bug tracker, and
review requests with a repository and bug tracker.

Reviewed at https://reviews.reviewboard.org/r/8904/