Do not pass user= twice in AvatarService.get_configuration_form
This unit test was not updated when the changes for avatar settings
forms went in. This patch updates the test to no longer user keyword
arguments for the fields that the base `ConfigForm` expects as to not
provide duplicate keyword arguments.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8928/