Fix upgrading the errordocs directory in rb-site.
rb-site was attempting to create a new errordocs directory for the new
500.html page. However, this actually failed when upgrading, since there
was already an errordocs symlink which then got removed after attempting
to create the directory. This meant that when we tried to write to the
new directory, it failed. A second upgrade attempt would work, since the
symlink was gone and the directory creation worked.

We now attempt the directory creation after removing the legacy symlink,
ensuring that upgrades work right the first time.

Testing Done:
Upgraded an existing site with the symlink. The upgrade worked.

Reviewed at https://reviews.reviewboard.org/r/8925/