Add support for touch events for comment creation in the diff viewer.
This adds touchscreen support in the diff viewer for selecting rows and
creating comments, and for opening existing comments. When using a
device like an iPad, users can now tap-and-drag down line numbers to
create a new comment, or tap a line with an existing comment to view it.
This works just like using a mouse.

The page will not scroll when creating a selection, helping to keep
selections predictable. Dragging in any other column (such as the code
columns) will continue to scroll the page.

Selections can be cancelled by ending a drag outside of the line numbers
column.

Testing Done:
Tested creating single-line comments on an iPad.

Tested creating multi-line comments on an iPad.

Tested dragging upwards and downwards, ensuring the selection is always
correct.

Tested ending a selection outside of the line numbers column to cancel
the selection.

Tested tapping a line with an existing comment to view that comment.

Tested scrolling the page by dragging in other parts of the page.

Tested that all mouse operations still worked correctly.

Reviewed at https://reviews.reviewboard.org/r/8930/