Add to_submitter_only to ReviewPublishedEmailHook
The `ReviewPublishedEmailHook` was missing the `to_submitter_only`
argument and we weren't sending it when filtering. This patch adds it
in. This patch also corrects the documentation of the
`ReviewpublishedEmailHook.get_cc_field` method to correctly refer to the
`cc_field` argument instead of `to_field`.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8894/