Deprecate the type argument of review request signals
This patch deprecates the `type` argument of the `review_request_closed`
and `review_request_closing` signals, due to shadowing the `type`
builtin. We still provide the `type` argument, but it is wrapped in a
lazy object that will emit a warning when casted to a unicode string. We
now include the `close_type` argument which takes the place of the old
`type` argument. Usages and unit tests have been updated for this new
behaviour.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8896/