Modernize hostingsvcs.service.URLRequest
The `URLRequest` initializer was using a mutable object (`{}`) for its
default value of its headers parameter, which would persist to other
requests if it was modified. We now use a default value of `None` and
change it to an empty dict in the method call.

Documentation has been added for the class and its methods.

Testing Done:
Built the docs and read through them.
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8909/