Fix a regression with the Quick Search functionality.
Some work was done in a recent release to fix issues in Quick Search and
to document the parameters going to this API. The addition of the
documentation (using a decorator that specifies the available list of
parameters) was using strict validation, not allowing for unknown
parameters. This broke Quick Search, which passes in a timestamp
argument (not otherwise used by this API) in all requests.

We now ignore any arguments that aren't used by the API, fixing this
regression.

Testing Done:
Unit tests pass.

Tested that Quick Search was working properly here. Sent the patch to
a customer and confirmed it fixed the problem for them.

Reviewed at https://reviews.reviewboard.org/r/8942/