Add an infobox for review requests.
This change introduces a new infobox for review requests, which is displayed
when hovering over links in the dashboard, user-supplied text (links to /r/N/),
or depends-on/blocks field contents. The idea is similar to existing infoboxes
for users and bugs-- it shows some information about the target, including the
submitter name, summary and description, issue counts, and some of the recent
reviews. This will allow users to get a quick overview of the review request
without having to actually open the page. In particular, this should hopefully
be most useful to see what the recent activity was.

Testing done:
- Code has been tested by running the development server and adjusting the tester
  review request for different scenarios.
- Ran Python and JavaScript unit tests.

Reviewed at https://reviews.reviewboard.org/r/8661/