Update the rendered body text of reviews when the models change.
The text shown for the body top/bottom fields on a review are no longer
fully static. When an API call is performed and the fields change in the
model, the review will re-render those fields.

To make this work, the ReviewModel resource class now tracks
HTML-rendered fields, along with the Markdown and raw fields, and the
ReviewView ensures that HTML is an included text type in responses. The
view then listens for changes on the raw fields and updates the page
based on the HTML fields.

Unit tests have been added to ensure this all works properly.

Testing Done:
Unit tests pass.

Tested along with an in-progress change and saw that the body text had
properly updated on model changes.

Reviewed at https://reviews.reviewboard.org/r/8993/