Add the PyMdown Markdown extension and enable strikethrough/subscripts.
PyMdown is a collection of useful Python-Markdown extensions, providing
all sorts of advanced functionality. We don't currently need or want the
vast majority of it, but it ships an extension for adding strikethroughs
and subscripts. While not an official Markdown feature, these are common
in many third-party Markdown parsers (and is rendered as such in
CodeMirror).

For some upcoming work, strikethroughs/subscripts would be very useful
to have. Down the road, we'll make use of a couple of other extensions
shipped in this package as well.

Testing Done:
Tested that strikethroughs and subscripts worked.

Reviewed at https://reviews.reviewboard.org/r/8991/