Support on-the-fly search indexing
On-the-fly search indexing is now supported for the Elasticsearch search
backend. When the option is enabled (from the search settings
administration form), the search index will be updated whenever review
requests are updated or deleted.

Additionally, the default value of the Whoosh search index has been
renamed to "search-index", which was the original default; it was
accidentally changed to "search_index" in a previous change.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8601/