Fix a security issue with querying comments when showing diff fragments.
The view that handles rendering comment diff fragments had an old query
that wasn't checking for access rights on the comment. This allows a URL
to be constructed that could fetch commented sections of diffs from
other review requests, even on other Local Sites.

This change bullet-proofs the query and adds several unit tests to cover
various access scenarios to ensure that diffs can only ever be accessed
when a user has the right permissions and when the comments pertain to
the review request being viewed.

This will be publicly disclosed at release time on Monday, June 19, 2017
and is embargoed until then.

Testing Done:
Unit tests pass.

Tested all possible scenarios manually, ensuring that comments can only
be viewed if the user has access rights to the review request and review
in question and that the comment belongs to the review request.

Reviewed at https://reviews.reviewboard.org/r/9019/