Prevent tracebacks for file not found errors in diff fragments views.
The diff fragments view logs a traceback when it hits an error, but we
don't always want this. If the review request exists but the files
backing its diff are now gone, every access would result in a very
lengthy traceback.

We no longer log File Not Found errors, instead just displaying the File
Not Found on the page instead. This will cut down a lot on noisy logs
for review requests like this.

Testing Done:
Tested with and without File Not Found errors and other exceptions.
Verified that we weren't getting log spam from File Not Founds.

Reviewed at https://reviews.reviewboard.org/r/9047/