Fix updating the RelatedObjectSelectorView with single-select views
This fixes an issue where the `RB.RelatedObjectSelectorView` would not
update its value due to a typo. The typo has been fixed and the object
in question has been replaced with a Map which is better suited to
mapping integer IDs to objects.

Testing Done:
Successfully was able to select a single user with this patch applied.

Reviewed at https://reviews.reviewboard.org/r/9036/