Fix linking to the diff index header on the Diff page.
The logic for building the review request contexts and links for the
Diff page was cleaned up a while back, but in the process, the link
generated to the diff file index was incorrect. The anchor name is
called "index_header," but the rewritten code attempted to use
"index-header". This updates the link to use the correct name.

Testing Done:
Tested that clicking Diff once again took me to the diff file index.

Reviewed at https://reviews.reviewboard.org/r/9039/