Move the user infobox view into accounts/ and make it class-based.
The user infobox has lived in reviews/ for a while now, but it's not
specific to review requests. The accounts/ module is a better place for
it. This change moves it there and also converts it to use class-based
generic views, making use of the new mixins offered by Djblets and
Review Board for access restrictions.

Testing Done:
Tested that the view still works, obeys access restrictions, and handles
ETag caching.

Reviewed at https://reviews.reviewboard.org/r/9000/